radiology utils and tests
