+++ data_processing/radiology/refined_table/annotation/generate.py
@@ -11,11 +11,6 @@ The PNG table contains paths to dicom png and overlay image that combines dicom 
 
 import os, time, shutil
 import click
-import numpy as np
-from PIL import Image 
-from medpy.io import load
-from filehash import FileHash
-from io import BytesIO
 
 from data_processing.common.CodeTimer import CodeTimer
 from data_processing.common.utils import generate_uuid_binary
@@ -23,97 +18,13 @@ from data_processing.common.config import ConfigSet
 from data_processing.common.sparksession import SparkConfig
 from data_processing.common.custom_logger import init_logger
 import data_processing.common.constants as const
+from data_processing.radiology.common.utils import overlay_images, create_seg_images
 
 from pyspark.sql import functions as F
 from pyspark.sql.types import StringType, IntegerType, ArrayType, StructType, StructField, BinaryType
 
 logger = init_logger()
 logger.info("Starting data_processing.radiology.refined_table.annotation.generate")
-
-
-def create_dicom_png(src_path, width, height):
-    """
-    Create a png from src_path image, and return image as bytes.
-    """
-    file_path = src_path.split(':')[-1]
-
-    data, header = load(file_path)
-
-    # Convert 2d image to float to avoid overflow or underflow losses.
-    # Transpose to get the preserve x, y coordinates.
-    image_2d = data[:,:,0].astype(float).T
-
-    # Rescaling grey scale between 0-255
-    image_2d_scaled = (np.maximum(image_2d, 0) / image_2d.max()) * 255.0
-
-    # Convert to uint
-    image_2d_scaled = np.uint8(image_2d_scaled)
-
-    im = Image.fromarray(image_2d_scaled)
-    # resize pngs to user provided width/height
-    im = im.resize( (width, height) )
-    
-    return im.tobytes()
-       
-
-def create_seg_png(src_path, uuid, width, height):
-    """
-    Create pngs from src_path image.
-    Returns an array of (instance_number, png binary) tuple.
-    """
-
-    file_path = src_path.split(':')[-1]
-    data, header = load(file_path)
-
-    num_images = data.shape[2]
-
-    # Find the annotated slices with 3d segmentation.
-    # Some reverse engineering.. save the instance numbers 
-    # from the series to identify the dicom slices that were annotated.
-    slices = []
-    for i in range(num_images):
-        image_slice = data[:,:,i]
-        if np.any(image_slice):
-            image_2d = image_slice.astype(float).T
-            # double check that subtracting is needed for all.
-            slice_num = num_images - (i+1)
-
-            image_2d_scaled = (np.maximum(image_2d, 0) / image_2d.max()) * 255.0
-            image_2d_scaled = np.uint8(image_2d_scaled)
-
-            im = Image.fromarray(image_2d_scaled)
-            # resize pngs to user provided width/height
-            im = im.resize( (int(width), int(height)) )
-
-            # save segmentation in red color.
-            rgb = im.convert('RGB')
-            red_channel = rgb.getdata(0)
-            rgb.putdata(red_channel)
-            png_binary = rgb.tobytes()
-
-            slices.append( (slice_num, uuid, png_binary) )
-
-    return slices
-
- 
-def overlay_pngs(instance_number, dicom_path, seg, width, height):
-    """
-    Create dicom png images.
-    Blend dicom and segmentation images with 7:3 ratio. 
-    Returns tuple of binaries to the combined image.
-    """
-    width, height = int(width), int(height)
-    dicom_binary = create_dicom_png(dicom_path, width, height)
-    
-    # load dicom and seg images from bytes
-    dcm_img = Image.frombytes("L", (width, height), bytes(dicom_binary))
-    dcm_img = dcm_img.convert("RGB")
-    seg_img = Image.frombytes("RGB", (width, height), bytes(seg))
-
-    res = Image.blend(dcm_img, seg_img, 0.3)
-    overlay = res.tobytes()
-
-    return (dicom_binary, overlay)
 
 
 @click.command()
@@ -178,7 +89,7 @@ def generate_png_tables(cfg):
         seg_png_table_path = os.path.join(project_path, const.TABLE_DIR, const.TABLE_NAME(cfg))
         
         # find images with tumor
-        create_seg_png_udf = F.udf(create_seg_png, ArrayType(StructType(
+        create_seg_png_udf = F.udf(create_seg_images, ArrayType(StructType(
                                     [StructField("instance_number", IntegerType()),
                                      StructField("scan_annotation_record_uuid", StringType()),
 				     StructField("seg_png", BinaryType())])))
@@ -203,10 +114,10 @@ def generate_png_tables(cfg):
         
         seg_df = seg_df.join(dicom_df, cond)
 
-        overlay_png_udf = F.udf(overlay_pngs, StructType([StructField("dicom", BinaryType()), StructField("overlay", BinaryType())]))
+        overlay_image_udf = F.udf(overlay_images, StructType([StructField("dicom", BinaryType()), StructField("overlay", BinaryType())]))
 
         seg_df = seg_df.withColumn("dicom_overlay",
-            F.lit(overlay_png_udf("instance_number", "path", "seg_png", F.lit(width), F.lit(height))))
+            F.lit(overlay_image_udf("path", "seg_png", F.lit(width), F.lit(height))))
  
         # unpack dicom_overlay struct into 2 columns
         seg_df = seg_df.select(F.col("dicom_overlay.dicom").alias("dicom"), F.col("dicom_overlay.overlay").alias("overlay"),
